+++ launch/giskardpy_pr2_mujoco.launch
@@ -1,8 +1,6 @@
 <launch>
  
-  <node pkg="giskardpy" type="giskard.py" name="giskard" output="screen">
-    <param name="config" value="PR2_Mujoco"/>
-  </node>
+  <node pkg="giskardpy" type="pr2_mujoco_open_loop.py" name="giskard" output="screen"/>
 
   <node pkg="giskardpy" type="interactive_marker.py" name="giskard_interactive_marker" output="screen">
     <rosparam param="enable_self_collision">False</rosparam>
@@ -15,4 +13,3 @@
   </node>
  
 </launch>
-
