+++ src/giskardpy/exceptions.py
@@ -4,7 +4,12 @@ from typing import Optional
 class DontPrintStackTrace:
     pass
 
+
 class GiskardException(Exception):
+    pass
+
+
+class SetupException(GiskardException):
     pass
 
 
@@ -108,6 +113,7 @@ class UnreachableException(PlanningException):
 class SelfCollisionViolatedException(PlanningException):
     pass
 
+
 # errors during execution
 # int64 EXECUTION_ERROR # if no execution code fits
 # int64 PREEMPTED # goal got canceled via action server interface
@@ -155,6 +161,7 @@ class ExecutionSucceededPrematurely(ExecutionException):
 
 class BehaviorTreeException(GiskardException):
     pass
+
 
 # ----------------------------------------------------------------------------------------------------------------------
 class ImplementationException(GiskardException):
